requirements.txt
alembic==1.0.10
aniso8601==7.0.0
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
colorama==0.4.1
factory-boy==2.12.0
Faker==1.0.7
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Mako==1.0.12
MarkupSafe==1.1.1
marshmallow==3.0.0rc7
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==4.6.3
pytest-flask-sqlalchemy==1.0.2
pytest-mock==1.10.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
text-unidecode==1.2
wcwidth==0.1.7
webargs==5.3.1
Werkzeug==0.15.4
zipp==0.5.1
